test: decouple metadata e2e test
