+++ packages/client/package.json
@@ -4,6 +4,7 @@
     "test": "npm-run-all test:**",
     "test:l1": "hardhat test ./test/arb.e2e.spec.ts --network localhost",
     "test:db": "hardhat test ./test/db.e2e.spec.ts --network localhost",
+    "test:metadata": "hardhat test ./test/metadata.e2e.spec.ts --network localhost",
     "start:read": "ts-node src/read.ts",
     "start:write": "ts-node src/write.ts",
     "lint": "eslint . --ext .ts --fix"
